chore: update create out point function
